Update... and hope you like it!
I try to do my best improve what author want, than i used Functions and Subroutine, and I put the code to create the text file and ask for user to create the folder if does not exsit. (I made this way beacuse the adm rights some time doesnot allow the user to create a folder on root address). don´t forget to put your name on that.